# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 2.0.0
  description: TransportPro - Transportation Business Management API
servers:
  - url: /api
...
paths:
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List clients
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/transactions:
    get:
      operationId: getClientTransactions
      tags: [clients]
      summary: Get client ledger transactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientLedger"
    post:
      operationId: createClientTransaction
      tags: [clients]
      summary: Add client transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientTransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientTransaction"
  /client-transactions/{id}:
    put:
      operationId: updateClientTransaction
      tags: [clients]
      summary: Update client transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientTransactionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientTransaction"
    delete:
      operationId: deleteClientTransaction
      tags: [clients]
      summary: Delete client transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        agoShortChargeRate:
          type: number
        pmsShortChargeRate:
          type: number
        notes:
          type: string
          nullable: true
        balance:
          type: number
          description: Outstanding balance (positive = owes us)
        createdAt:
          type: string
      required: [id, name, agoShortChargeRate, pmsShortChargeRate, balance, createdAt]
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            transactions:
              type: array
              items:
                $ref: "#/components/schemas/ClientTransaction"
          required: [transactions]
    ClientInput:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        agoShortChargeRate:
          type: number
        pmsShortChargeRate:
          type: number
        notes:
          type: string
          nullable: true
      required: [name, agoShortChargeRate, pmsShortChargeRate]
    ClientTransaction:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        type:
          type: string
          enum: [invoice, advance, payment, adjustment]
        amount:
          type: number
        reference:
          type: string
          nullable: true
        batchId:
          type: integer
          nullable: true
        batchName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        transactionDate:
          type: string
        createdAt:
          type: string
      required: [id, clientId, type, amount, transactionDate, createdAt]
    ClientTransactionInput:
      type: object
      properties:
        type:
          type: string
          enum: [invoice, advance, payment, adjustment]
        amount:
          type: number
        reference:
          type: string
          nullable: true
        batchId:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        transactionDate:
          type: string
      required: [type, amount, transactionDate]
    ClientLedger:
      type: object
      properties:
        client:
          $ref: "#/components/schemas/Client"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/ClientTransaction"
        totalInvoiced:
          type: number
        totalAdvances:
          type: number
        totalPayments:
          type: number
        balance:
          type: number
      required: [client, transactions, totalInvoiced, totalAdvances, totalPayments, balance]
...